@@ -1,17 +1,17 @@
 name: flutter_nucleus
 description: Use the atomic state management package nucleus with Flutter.
-version: 0.3.5
+version: 0.4.0
 homepage: https://github.com/tim-smart/nucleus
 
 environment:
-  sdk: ">=2.17.0 <4.0.0"
+  sdk: ">=3.0.0 <4.0.0"
   flutter: ">=1.17.0"
 
 dependencies:
   flutter:
     sdk: flutter
   flutter_hooks: ">=0.18.0 <1.0.0"
-  nucleus: ^0.3.8
+  nucleus: ^0.4.0
   shared_preferences: ">=2.0.0 <3.0.0"
 
 dev_dependencies:
